[feat] : Enum <-> String 간 변환 메서드 추가
