--- core/src/main/java/dev/handsup/auction/domain/product/product_category/ProductCategoryValue.java
@@ -1,14 +1,0 @@
-package dev.handsup.auction.domain.product.product_category;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ProductCategoryValue {
-	DIGITAL("디지털 기기"),
-	FURNITURE("가구"),
-	FASHION("패션");
-
-	private final String description;
-}
