--- core/src/main/java/dev/handsup/exception/ErrorCode.java
@@ -1,8 +1,0 @@
-package dev.handsup.exception;
-
-public interface ErrorCode {
-
-	String getMessage();
-
-	String getCode();
-}
