--- core/src/main/java/dev/handsup/exception/ValidationException.java
@@ -1,14 +1,0 @@
-package dev.handsup.exception;
-
-import lombok.Getter;
-
-@Getter
-public class ValidationException extends RuntimeException {
-
-	private final String code;
-
-	public ValidationException(ErrorCode errorCode) {
-		super(errorCode.getMessage());
-		this.code = errorCode.getCode();
-	}
-}
